+++ board/RPXClassic/RPXClassic.c
@@ -218,7 +218,7 @@ static long int dram_size (long int mamr_value, long int *base, long int maxsize
 
 	memctl->memc_mamr = mamr_value;
 
-	return (get_ram_size(base, maxsize));
+	return (long int)(get_ram_size((unsigned long)base, (unsigned long)maxsize));
 }
 /*-----------------------------------------------------------------------------
  * aschex_to_byte --
